+++ examples/tag_example/main.go
@@ -36,30 +36,7 @@ func tagExample() (err error) {
 		"Warn log",
 	)
 
-	// Retrieve our standard logger
-	stdLogger := log.GetLogger("std")
-
-	// Change the tag colors!
-	// You can change other text attributes too, such as bold, underline and italics
-	stdLogger.SetTagColorMap(log.TagColorMap{
-		// No entry for #tag1 so it prints Faint FgWhite (grey) by default
-		"tag2": []log.Attribute{log.FgMagenta},
-		"tag3": []log.Attribute{log.FgCyan},
-	})
-
-	// Write to all loggers one more time
-	log.TError(
-		log.Tags{"tag1"},
-		errors.New("Error log"),
-	)
-	log.TInfo(
-		log.Tags{"tag1", "tag2"},
-		"Info log",
-	)
-	log.TWarn(
-		log.Tags{"tag2", "tag3"},
-		"Warn log",
-	)
+	// If you want to change the color of your tags from the default (grey), see ../color_example/main.go
 
 	return
 }